// frontend/src/shared/api/generated/models/eventRatingResponse.ts
/**
 * Generated by orval v8.19.0 🍺
 * Do not edit manually.
 * CodigoActivo.API
 * OpenAPI spec version: 1.0
 */

export interface EventRatingResponse {
  id?: string;
  eventId?: string;
  userId?: string;
  score?: number;
  /** @nullable */
  mostLiked?: string | null;
  /** @nullable */
  leastLiked?: string | null;
  /** @nullable */
  suggestions?: string | null;
  createdAt?: string;
  /** @nullable */
  updatedAt?: string | null;
}