// frontend/src/shared/api/generated/models/saveEventRatingRequest.ts
/**
 * Generated by orval v8.19.0 🍺
 * Do not edit manually.
 * CodigoActivo.API
 * OpenAPI spec version: 1.0
 */

export interface SaveEventRatingRequest {
  /** @nullable */
  score?: number | null;
  /** @nullable */
  mostLiked?: string | null;
  /** @nullable */
  leastLiked?: string | null;
  /** @nullable */
  suggestions?: string | null;
}